Remove assertions from core, confine assertj in test scope
