+++ unittest-support/src/main/java/org/seedstack/seed/test/rule/MockitoRule.java
@@ -7,7 +7,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.seed.unittest.helper;
+package org.seedstack.seed.test.rule;
 
 import org.junit.rules.TestRule;
 import org.junit.runner.Description;
